# api/openapi.yaml
openapi: 3.0.3
info:
  title: Tab Memory API
  description: |
    REST API for Tab Memory Chrome Extension - manage browsing sessions, tab events, and user data.
    
    ## Authentication
    All endpoints except `/auth/*` require a Bearer token in the Authorization header.
    
    ## Rate Limiting
    - 100 requests per 15-minute window per IP
    - Authenticated requests have higher limits
    
    ## Data Flow
    1. Register/login to get access token
    2. Extension sends tab events to `/events/batch`
    3. API processes events and creates sessions
    4. Use `/sessions` endpoints to retrieve and restore sessions
  version: 1.0.0
  contact:
    name: Tab Memory API
    url: https://github.com/your-org/tab-memory
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://tab-memory-api.onrender.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from `/auth/login` or `/auth/register`
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
        email:
          type: string
          format: email
          description: User email address
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
      required:
        - id
        - email
        - created_at
...
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
        refresh_token:
          type: string
          description: JWT refresh token
        expires_in:
          type: integer
          description: Token expiry time in seconds
        user:
          $ref: '#/components/schemas/User'
      required:
        - access_token
        - refresh_token
        - expires_in
        - user
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the request succeeded
        data:
          type: object
          description: Response data (varies by endpoint)
        error:
          type: string
          description: Error message if success is false
        message:
          type: string
          description: Additional information
      required:
        - success
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Human-readable error message
        stack:
          type: string
          description: Stack trace (development only)
      required:
        - success
        - error
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      summary: Register new user
      description: Create a new user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                  maxLength: 128
              required:
                - email
                - password
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthResponse'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      summary: User login
      description: Authenticate user and return access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required:
                - email
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      summary: Refresh access token
      description: Get new access token using refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
              required:
                - refresh_token
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          access_token:
                            type: string
                          expires_in:
                            type: integer
...